# Ix: Improve Min(), MinBy(), Max(), MaxBy() performance