--- a/Ix.NET/Source/Benchmarks.System.Interactive/Program.cs
+++ b/Ix.NET/Source/Benchmarks.System.Interactive/Program.cs
@@ -19,6 +19,7 @@ namespace Benchmarks.System.Interactive
             var switcher = new BenchmarkSwitcher(new[] {
                 typeof(BufferCountBenchmark),
                 typeof(DeferBenchmark),
+                typeof(MinMaxBenchmark),
             });
 
             switcher.Run();